// src/logistics-api/model/updateDriverStatusByDriver.ts
/**
 * Logistics API
 * Logistics API provides services to find and manage entities for logistics
 *
 * OpenAPI spec version: 1.0
 * Contact: vladislavprokopenko1@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Object representing Driver status updating dy Driver
 */
export interface UpdateDriverStatusByDriver { 
    /**
     * rest/driving status
     */
    status?: UpdateDriverStatusByDriver.StatusEnum;
}
export namespace UpdateDriverStatusByDriver {
    export type StatusEnum = 'REST' | 'DRIVING';
    export const StatusEnum = {
        REST: 'REST' as StatusEnum,
        DRIVING: 'DRIVING' as StatusEnum
    };
}